chore(front): rename FileCard to FileCardOrg that will be deleted when front is standardised
